// src/types/events.ts
export enum Events {
    LOAD_LOTS = 'catalog:changed',
    OPEN_LOT = 'card:open',
    OPEN_BASKET = 'basket:open',
    CHANGE_LOT_IN_BASKET = 'lot:changed',
    VALIDATE_ORDER = 'formErrors:changed',
    OPEN_FIRST_ORDER_PART = 'order_payment:open',
    FINISH_FIRST_ORDER_PART = 'order:submit',
    OPEN_SECOND_ORDER_PART = 'order_contacts:open',
    FINISH_SECOND_ORDER_PART = 'contacts:submit',
    PLACE_ORDER = 'order:post',
    SELECT_PAYMENT = 'payment:changed',
    INPUT_ORDER_ADDRESS = 'order.address:change',
    INPUT_ORDER_EMAIL = 'contacts.email:change',
    INPUT_ORDER_PHONE = 'contacts.phone:change',
    OPEN_MODAL = 'modal:open',
    CLOSE_MODAL = 'modal:close',
}